Implement simple data transformation DSL
So the goal for this project is to extract structural information about
composite components and process exports from the Babel AST.

All the information is there, we just need to transform it. So i'm
writing a simple object transformation DSL that we will use to
do that transformation.

It works with transform function:

    transform(node, rule)

The node is a AST node or a value within such a node and rule is a rule
object that determines what to do.

  - if it's true, copy that property from the node
  - if the rule value is a string, use that string
  - the rule value is a function, call the function with the node and
    use the result unless it is undefined

Furthermore, we make sure to call the same rule multiple times for
arrays and recurse into complex objects.

Helpers
=======

We implement two general helper functions

    Swith({ [key], ruleMap })

switches based on an object property (default "type") and choses the
value of the rule key that matches that property or the "default" key
if none match. If there is no "default" key the result is undefined and
will be ignored.

    Any(ruleMap)

works on maps and applies the given ruleMap to any key.

This commit also adds test-cases for the DSL.