Create full object identifiers for ObjectPatterns.
Instead of just taking the name of identifier nodes like we mostly do,
use an object with type and name for object patterns.

Seems bad to had the value sometimes be a string and sometimes be a
complex object. Feels more clean this way even though we should be able
to handle the other way with svenson, too.