Extract information about project exports.
Extracts two kinds of things from the process modules. Those are all
modules for which the module name is equal to the process name and that
are not in the /composites/ folder.

Extraction of initProcess data
==============================

The initprocess method configures process and/or scope and returns the
process states map.

it must be defined in the form

    export function initProcess(process,scope)
    {
        // ... CONFIG ...

        return (... STATE MAP ...)
    }

In the config area we record every property write and every method
call on the process and scope parameters.

The return value must be the proces map

Process Map
-----------

The process map must have a shape like this.

    {
        startState: "CustomerList",
        states: {
            "CustomerList": {
                "to-detail":
                    {
                        to: "CustomerDetail",
                        action: scope.addTodo
                    }
            },
            "CustomerDetail": {
                "save" : {
                    to: "CustomerList",
                    action: t => { process.back() }
                },
                "cancel" : {
                    to: "CustomerList"
                }
            }
        }
    }

`startState` contains the name of the initial state of the process.

`states`contains the state map which has one entry per state name.
Each state map entry in turn is a transisition map mapping transition
names to transition objects.

the objects can have two properties.

`to` contains the target state name of the transition. If the action
choses the target dynamically, the `to` property can be left out.

The `action` property can contain either a function expression or direct
scope action reference.

The function expression can receive a transition object, here called
`t`.

Extraction of the scope
=======================

We expect the default export of the process module to be the mobx
decorated scope implementation.

We record the name and extract four kinds data from the scope class.

Observable properties
---------------------

    @type('PagedFoo')
    @observable currentFoo = ...;

We extract our own type decoration which for now is basically just an
annotation doing nothing and the observable decoration which categorizes
this class property as "observable". The value can be any static
expression matching the type or the `injection` method to push GraphQL
results from the server during process init.

Actions
-------

    @action myAction(param)
    {
        ...
    }

Actions annotated with @action. We take the params as array and body as
code-block.

Computeds
---------

    @computed
    get computedProperty()
    {
        ...
    }

We record the computed blocks and take their body as code-block

Helpers
-------

We allow generic methods to be defined as helper methods of which we
take the params as array and the body as code-block