Webpack plugin: Write to many files
This changes our plugin to take two options, a appsPath option for the
source JavaScript files and a modelPath option for the output path in the
webapp folder, which will be /WEB-INF/automaton.

This writes all collected entries into the same hierarchy as the
Javascript files, only as JSON data and with JSON endings.

The JSON data is meant to flow back into the generator to be considered
by new generation runs that update the same JSON data.