Extract scopes from scopes.js
Extract data from a resource with the module pattern

  ./apps/<app-name>/scopes

Refactors out the analysis of the process scope and do the same for named
exports of that module.

Currently supported scope names are ApplicationScope, UserScope and
SessionScope, all are written to a decapitalized propery of the module
data if they exist.

Adds a test for scope data extraction.