Add support for analyzing domain models
We need support for client-side modeled domain objects. The objects can
be observable with our without a modeled class backing them (Mobx
handles JSON object maps just fine), but we need a class model to be
able to have local @computed and @action implementations.

This commit generalizes the existing Scope analysis and uses it to
extract the same information from the default exports of modules like

    ./apps/my-app/domain/MyDomainType.js

The default export must be named like the module.