Support discard and confirmation on transition
We extended our transition model.

The discard property can be set to `true` to make the transition happen
even if inputs are not valid for the current form. Buttons referencing
such transitions will be *not* disabled when there is a form error.

The confirmation property can contain a function that takes a future
transition context as argument and returns a truthy string to pop
up a confirmation displaying that string as question to the user.

Extracts both into our transition entries.

Adds a test.