Switch to new composite format
Following the general changes to a more hook / functional component
based standard for our components, we also define the composite format
to be that.

Hooks are just "constants" in our lingo, so now we could even support
life-cycle within the composite format.

We only accept one kind of React functional component

```js
    import React from "react"

    import { observer as fnObserver } from "mobx-react-lite"

    const SimpleComposite =
        /* must be named "props" */
        props => {
            return (
                <div>
                    <h1 className="test-class">SimpleComposite</h1>
                </div>
            )
    }

    export default fnObserver(SimpleComposite)
```

This commit changes the AST analysis to support above format.

Changes all composite test cases to the new format and changes the tests
to support the slightly changed format. There is no decorators anymore,
and instead we have an `fnObserver` HOC call on the export.

Refactors FormComposite away from the odd old form code.