Add helper properties
Our initial assumption that all properties of our scope will observable
properties proved incorrect. Concrete example is now a WorkingSet prop
that is not an observable as-such but just uses observables internally.

Adds a working set to the process test.

Makes it so that properties without @observable our now added to the
helpers. So helpers are now either a helper method as before with the
same syntax or a helper property with the JSON structure being the same
as for the observables.

Adjusts resulting test data.

Adds Helper properties to reference docs