Implement process v2 model extraction
Contains the necessary changes to extract the model data from source in
the new process format.

This follows the changes in automaton-js and the output of the
process-v2-codemod.js transformer.

We keep the composites for now because we still have some one those in
automaton-test and it also allows for less code changes.

We keep testing the composite features in composites even though those
play a lesser to no role now.

And then we just reuse composite parsing to extract data from our third
ViewState() paramater, the component function.

We move the new init code that decides on the start state to an init
array prop within the processExports which of course now loses its
"process" prop because that no longer exists.

if the return is a simple identifier return we copy the start state name
to the new startState prop, which is `null` for a complex start state
init code.

We change some of the tests for the new structure and rename one to
"extracts view state models" which is the only test for the new state
model so far.

The model is pretty simple though as it contains only already existing
data structures. The former stateMap is now reduced to being a
transitionMap within the states.